refactor: extract bundle curve construction
