+++ bpy_speckle/converter/from_bundle/_baking/geometry/__init__.py
@@ -1,0 +1,1 @@
+"""Geometry construction internals."""
